modified

+++ src/TC264_Samrt_Car_Demo_v2.2/TC264_Samrt_Car_Demo_v2.2/0_Src/AppSw/Tricore/Main/ServeSource.c
@@ -137,7 +137,9 @@ uint8 Bluetooth_Read_Data(void)
 }
 /*****************************************舵机电机相关************************************************************/
 //详见PwmDemo.h PwmDemo.c
-uint32 SetMotor(float direction, float pwmDuty)//电机操作函数 方向+占空比（%）
+
+// one 'float' has been changed into 'int'
+uint32 SetMotor(int direction, float pwmDuty)//电机操作函数 方向+占空比（%）
 {
 	if (direction == FORWARD)
 	{
@@ -151,8 +153,8 @@ uint32 SetMotor(float direction, float pwmDuty)//电机操作函数 方向+占空比（%）
 	}
 	return 0;
 }
-
-uint32 SetSteer(float direction, float pwmDuty)//舵机操作函数 方向+占空比（%）
+// one 'float' has been changed into 'int'
+uint32 SetSteer(int direction, float pwmDuty)//舵机操作函数 方向+占空比（%）
 {
 	if (direction == LEFT)
 	{
